fix NPE in sentence search adapter
